ADAIG-24: restructured folders
